refactor: convert code to use ECMAScript module syntax
